refactor, replicate for evm charts
